Move revision row to separate component
